requirements.txt
aiohttp
asyncpg
brotli
discord-ext-menus
discord.py
httpx
jishaku
levenshtein
lingua-language-detector
lxml
psutil
python-dateutil
recurrent
tldextract
toml
uvloop
